refactor(api): Reorder import statements, update interceptor logic
This commit reorders the import statements in the `index.ts` file of the `api` directory. The `useAuthStore` import statement is moved below the `useRouter` import statement to ensure consistent ordering.

Additionally, the logic of the `API.interceptors.request` callback function is updated to use the `authStore.userRole` property instead of a hardcoded role. If the user role is "檢視者", a `Promise` will be rejected with the error message "權限不足". This change improves the authorization flow in the application.

---

fix(component): Remove websocket logging and add prop

This commit removes the console logging statement in the `VideoCard.vue` component. The statement was logging the `event.data` value received through the websocket.

Additionally, a prop `video` is added to the component. This prop is required and should be of type `IPCam`. The prop is used to update progress information based on the video being processed.

---

feat(component): Add save modal and functionality

This commit adds a save modal to the `VideoScheduler.vue` component. The modal is triggered when the "Save" button is clicked. It displays a confirmation message and offers the options to confirm or cancel the save operation.

The save functionality is implemented with the `handleSave` method. When the save is confirmed, a success message is displayed and the modal is closed.

+++ src/api/index.ts
@@ -1,6 +1,6 @@
-import { useAuthStore } from '@/stores/auth';
 import axios, { AxiosRequestConfig, AxiosResponse } from 'axios';
 import { useRouter } from 'vue-router';
+import { useAuthStore } from '@/stores/auth';
 
 const router = useRouter();
 const API = axios.create({
@@ -9,6 +9,7 @@ const API = axios.create({
 		'Content-Type': 'application/json',
 	},
 });
+
 // 攔截器加入 Auth token
 API.interceptors.request.use((config) => {
 	const authStore = useAuthStore();
@@ -83,6 +84,11 @@ export const createIPCam = (imei: IPCam): Promise<void> => {
 };
 // 刪除 IPCam
 export const deleteIPCam = (imei: IPCam): Promise<void> => {
+const authStore = useAuthStore();
+
+	if (authStore.userRole === '檢視者'){
+		return Promise.reject('權限不足')
+	}
 	return API.delete(`/ipcams/${imei}`);
 };
 // 獲取指定時間段內的截圖列表
@@ -98,6 +104,10 @@ export const listSnapshotsInRange = async (imei: string, startTime: number, endT
 
 // 上傳一張照片
 export const createSnapshot = (ipcamImei: string, image: File, createdAt?: number): Promise<void> => {
+	const authStore = useAuthStore();
+	if (authStore.userRole === '檢視者'){
+		return Promise.reject('權限不足')
+	}
 	const formData = new FormData();
 	formData.append('ipcam_imei', ipcamImei);
 	if (createdAt) {
@@ -131,6 +141,11 @@ export const downloadSnapshotById = async (snapshotId: number) => {
 
 // 刪除照片
 export const deleteSnapshotById = (snapshotId: number) => {
+const authStore = useAuthStore();
+
+	if (authStore.userRole === '檢視者'){
+		return Promise.reject('權限不足')
+	}
 	return API.delete<void>(`/snapshots/${snapshotId}`);
 };
 // 下載照片的縮圖
@@ -161,16 +176,28 @@ export const getAllVideos = () => {
 
 // 請求建立影片
 export const createVideo = (videoInput: VideoInput) => {
+	const authStore = useAuthStore();
+	if (authStore.userRole === '檢視者'){
+		return Promise.reject('權限不足')
+	}
 	return API.post<void>('/videos', videoInput);
 };
 
 // 刪除影片
 export const deleteVideo = (videoId: number) => {
+	const authStore = useAuthStore();
+	if (authStore.userRole === '檢視者'){
+		return Promise.reject('權限不足')
+	}
 	return API.delete<void>(`/videos/${videoId}`);
 };
 
 // 取消製作中的影片
 export const cancelCreateVideo = (videoId: number) => {
+	const authStore = useAuthStore();
+	if (authStore.userRole === '檢視者'){
+		return Promise.reject('權限不足')
+	}
 	return API.delete<void>(`/videos/cancel/${videoId}`);
 };
